refactor(models, dao): Remove `Branch`, `Property`, and `Warehouse` classes and DAOs, and adjust package structure
- Deleted `Branch`, `Property`, and `Warehouse` classes alongside related DAOs.
- Renamed package `models` to `modelos` for consistency.
- Updated all references and `module-info.java` to reflect the new package structure.
