+++ src/main/java/mx/edu/utez/warehousemanagerfx/modelos/Property.java
@@ -1,4 +1,4 @@
-package mx.edu.utez.warehousemanagerfx.models;
+package mx.edu.utez.warehousemanagerfx.modelos;
 
 public class Property {
     // Attributes
